requirements.txt
pydantic
fastapi
uvicorn[standard]
requests
sqlmodel
psycopg[binary]
python-dotenv
python-jose
bcrypt
python-multipart
langchain
langchain-core
langchain-openai
langchain-anthropic
langgraph
langgraph-supervisor
pinecone
langchain-pinecone
langchain-tavily
mem0ai
pyyaml
langgraph-checkpoint-postgres
python-dotenv